refactor signup functions
